* Moved python/__init__.py python/packagekit/__init__.py
* Moved backends/yum/helpers/packagekit.py -> python/packagekit/backend.py
So that the PackageKitBaseBackend python class used byt yum & conary is part
of the packagekit python module and be used by
'from packagekit.backend import PackageKitBaseBackend'
added links from python/packagekit  to yum/helpers & conary/helpers so
that the python code will work when started in the directory.
make changes to the Makefile's to support the changes.