Adds a boolean flag recursive to the GetDepends and GetRequires method.
>From the description in the developer documentation:

GetRequires:
If recursive is true: This means if gnome-power-manager depends on
NetworkManager and NetworkManager depends on HAL, then GetRequires on
HAL should return both gnome-power-manager and NetworkManager.

GetDepends:
If recursive is true: This means if gnome-power-manager depends on
NetworkManager and NetworkManager depends on HAL, then GetDepends on
gnome-power-manager should return both HAL and NetworkManager.

Why does this belong in the backends? Well, I tried doing this with lots
of recursive calls to GetRequires(package_id) and for gimp it took a
long time, when in yum took a few seconds. The backend doesn't have to
do all this IPC, and so can do it much quicker.

If your backend can't do a recursive GetDepends or GetRequires, just
ignore the flag; anything is better than nothing.