Added a new filter 'basename' to the list of allowed filter types and use with GetUpdate.
The basename filter will only return results according to the package basename.
This is useful if you are searching for 'pm-utils' and you only want to show the
main pm-utils package, not any of the -devel or -debuginfo or -common suffixes
in the UI. The basename is normally the original name of the srpm or source deb.

This parameter is valid when searching, but it will be most useful when it is
used with GetUpdates, which I've added it to.