Add a new DBUS interface org.freedesktop.PackageKit.Notify so we can get system wide notifications from PackageKit, without having to bodge PkClient to do this for us.
As a general note, PackageKit abuses the concept of DBUS paths, having all the
transactions on on path, and doing the demux itself. This patch is the first
of many to make the DBUS interface a whole lot more sane.

This was triggered by someone pointing out why this comment existed:
/* we always emit, even if the tid does not match */