yum: add a new untested class to cache the comps list
Getting the comps list takes me 600ms -- which is unacceptable when we are using this
interactivly.
Create a sqlite cache in /var/cache/yum/packagekit-groups.sqlite so we can get
the groups list for any package in less than 5ms.

Ultmately this code should be moved to yum, although we'll need some heavy testng before this