apt: handle inconsistent source repo state.
If deb-src lines aren't available for every deb line then
softwareproperties will report it as inconsistent, with
None. The apt backend tried to pass this directly as
the "enabled" parameter, which fails as it should be able to
be considered an integer. We instead substitute False.

Thanks to Baptiste Mille-Mathias for reporting and testing.