bugfix: execute the pk_transaction_run handler in an idle callback
After spending a couple of hours reviewing the code, it was obvious that queued transactions were
being called recursively, which really screwed up the next transaction list scheduling.
This would often leave the daemon 'waiting for transactions' indefinately.
Make sure we idle_add the running callback, so we break out of the signal chain and add a metric ton of
self tests to verify this is working correctly.