bugfix: when we have more than one transaction queued, don't try to run them all at once
This bug was never triggered before as we were running eachpk_transaction_run in a loop
and so blocked at this point. Now we are idle_add'ing them, it's possible for all of
the remaining queue to be dispatched at the same time.
This fixes the bug where the dispatcher would sometimes fail to run the second method
and PkSpawn would error out with 'trying to set timeout when already set'.