bugfix: fix a theoretical hang where we could try to reuse a exiting instance
As pk_spawn_exit() was blocked in pk_spawn_argv(), a new process could not be spawned when
the instance was shutting down. This was not the case when pk_spawn_exit() was called from
PkBackendSpawn, and so if the backend was in the process of exiting and a new command was
send to the dispatcher, the daemon would stop processing existing requests.
The fix is quite simple, forcing pk_spawn_exit() to wait for completion in all cases,
and then to add an extra flag to differentiate between a forced exit and the changing of
the dispatcher