bugfix: fix protocol violation in the server and client
PackageKit does not adhere to it's own specification where a new TID is required
when an error is emitted on the original TID. For a couple of releases now
we cleanup unused TID's after 5 seconds, to avoid denial of service attacks.
As we re-use the original TID after showing the signature dialog, we wait on
the client, and then resumbit after PolicyKit returns from the authentication
request.
If the user takes longer than 5 seconds to authenticate, then the TID is timed
out, and the helpful error is presented to the user:

Method 'Update System' with signature '' on interface
'org.freedesktop.PackageKit.Transaction' doesn't exist.

Of course, the correct fix is to release the TID on error to prevent it being
reused, and also to re-request a TID after authentication has been obtained.