feature: remove the 'desktopnames' parameter from the x-packagekit-plugin code and add the desktop icon if installed
Specifying a desktop file would allow the user to launch a different application to the one they specified
as the package. This would be bad. We can use the pk_desktop_get_shown_for_package() api call to get the
best desktop file, and so we can use the icon in this case.