bugfix: Check the sender for SetLocale and the UID for Cancel
For a long time other users were able to cancel each others jobs.
On a multiuser workstation this is not a good idea and is now
controlled using PolicyKit.

Another issue was that a different connection could set the locale
from the connection that asked for the TID. We should always verify
that the sender is the same, as this prevents a local DOS on a
multiuser system.