Trivial patch to fix the python frontend.
The old one was probably for an old version of PackageKit, or
just flat out broken.  This patch fixed the frontend to comply
with the correct DBUS signatures and signals as found on
http://www.packagekit.org/gtk-doc/Transaction.html.

Also added PolicyKit authorization.

To use, extend the PackageKit class and override the callbacks
and have them do what you want them to do.

Signed-off-by: Thomas Liu <tliu@redhat.com>