Initialize relevant GError's to NULL
Hi, I've been debugging a packagekit backend for my distribution (Arch Linux)
and came across a GLib warning about setting a GError twice. I did a grep for
uninitialized GError's and found lots in src/pk-transaction.c, which I assume
was why the warnings occured. Attached is a patch to fix this.

Signed-off-by: Richard Hughes <richard@hughsie.com>