urpmi: fix update-system command and only_trusted parameter
incorrectly modified two calls to perform_installation to pass options hash as a reference,
causing the parameters to not be interpreted correctly by
perform_installation, breaking update-system command and causing
only_trusted to be ignored in these cases.

Fix that by passing the options as key/value pairs as expected by
perform_installation.