yum: Fix update notices: not all yum repos are official fedora ones
Currently a hardcoded link pointing to https://admin.fedoraproject.org/updates/
is added to all update notes that have an update_id.  But not all repositories
providing updateinfo.xml data and id's are the upstream fedora repositories.

Ideally the fedora repos would provide the link themselves in a references
entity, and I'll add a bz for them to do that, but in the meantime the
following patch ensures the link only gets added to updates from the official
fedora repo by checking that the 'from' matches 'updates@feodraproject.org'.

I've also moved this logic before parsing other references as this
vendor-specific URL is likely the most important (and self) reference.

Signed-off-by: Richard Hughes <richard@hughsie.com>