Ensure we set an error if the spawned backend gets cancelled
We can't assume the spawned backend gets a chance to set an error when we SIGQUIT it
and so set a generic transaction cancelled error code when before we send the signal.

This means we don't trigger the obnoxious internal error when sometimes cancelling
tranasctions, e.g. '...This is a serious error as the spawned backend did not...' and
we don't scare the user when there's nothing actually wrong.