Add the concept of a transaction lifecycle
This new lifecycle removes the committed, running and finished booleans,
as keeping the bitfield sane was a waste of time when the lifecycle only
went one way, and only ever had one state at a time. This allows us to
make the code simpler, and makes it much easier to in future split up the
pre-running, running and post-running states.

Apologies for making your life harder Matthias.. :-)