trivial: Select KDE Debconf frontend if KDE is running
This has no effect on the Debconf process at time,
since the official Debconf-KDE frontend is not yet
finished.
If dialog frontend is "kde", but the frontend was not
found, Debconf will automatically fallback to "gnome"
or "dialog".

This patch also adds a few more temp files which
are left by the APTcc backend to the gitignore list.