zif: Add a new backend, which uses Zif to interface with the rpmdb and the Fedora metadata
This new backend should be considered experimental, and should not be used
in production environments. It is often an order of magnitude quicker then
the 'yum' backend as there is no need to spawn a process in python and
certainly no need to check the metadata on every access.

There will be bugs, so please test carefully. More information is at
https://github.com/hughsie/zif/raw/master/README.PackageKit for how
to switch to the new backend from yum.