Enable use of new callback mode on yum versions that support it
The new callback mode saves (tens of) megabytes of memory on
larger transactions as it avoids keeping rpm headers loaded in
memory needlessly. This enables the new mode on yum versions that
support it, on older versions it does nothing at all. As PK callback
is inherited from yum.rpmtrans.RPMBaseCallback, it's shielded from the
underlying changes so no further modifications are needed for the new
callback mode.

Signed-off-by: Richard Hughes <richard@hughsie.com>