zypp: fix deletion of package locks (bnc#690038)
Save and restore Pool status when doing calculations like
in get_depends and get_requires.

Fixes bug introduced by commit:
"zypp: reset packages status in repo while removing packages" which
resetted the status of an item, including its locks.

So if you browse a locked package, see if requires, the reset done
after removed the locks, and if you installed a package later, this was
saved.