Add an ItemProgress signal designed to eventually replace SubPercentage
Most GUI's try to link the SubPercentage with a package, in order to do sane
per-package completion reports. This is a hacky heuristic that just works around
a deficiency in the public API. The ItemProgress is a new signal that's designed
to be emitted by backends tying the package to the percentage and should make
all the evil hacks go away.

We've not going to drop SubPercentage just yet, as the client code will need
time to convert to the new world order.