Add a "Python" type to PkProvidesEnum and document how the provided components should be named.
Also renamed the "LIBRARY" type to "SHARED_LIB", as we don't
search for all libraries but only for shared libs.
(This change is possible because there hasn't been a PK release
with this change yet, so this doesn't break existing API)

The component naming shemes are documented in
docs/provides-component-naming.txt. This file is still incomplete
and needs to be extended.
Also, I'm not really happy with it's content (yet).