glib: catalog: Skip empty entries
When parsing the catalog file we use g_key_file_get_string() instead
of g_key_file_get_string_list(), and then split the string into parts.
For string lists, the last character is a ';' so if the user doesn't
know this, we add an empty string to the array, leading to errors
further on. Therefore, skip empty strings.

https://bugs.freedesktop.org/show_bug.cgi?id=43305

Signed-off-by: Richard Hughes <richard@hughsie.com>