Remove the only_trusted flag and Simulate* methods in favour of using transaction flags
NOTE: This commit breaks backend and client API. If the client is using PkTask
then only a small change is required, but using D-Bus or PkClient the client
will have to be modified substantially.

This change is designed to make the code much smaller and more maintainable,
and also so we can add features in the future without adding loads of extra API.
If you have any questions, please ask on the PackageKit mailing list.