glib: Don't update the transaction role to 'unknown'
If the proxy is being setup, the daemon transaction role may not have been set
to the correct value yet. Use the client-set role value until the daemon issues
a PropertyChanged signal which will update the role only if it's different to
the client-set role.

This means that the role should not change at all for the client in normal
circumstances.