Move the libpackagekit-qt code to a separate project
The QT code has a different ABI and API policy to the GLib code and it doesn't
make sense to keep them linked together. Also, splitting out the Qt code allows
us to use cmake as a build system which is a better fit for the Qt library.

It's expected that Daniel and Matthias will take over the new project, although
I'm sure I'll be committing to the new repo too to keep the interfaces up to
date.