Make the client-helper run on non-Linux systems
Apparently it's a bad idea to link relative from the test program,
since it might start up in a ".libs" directory under the current.
It's also a bad idea to assume that /usr/bin/python always exists,
for instance on FreeBSD it is optional and in /usr/local/bin/python.