systemd-updates: Do not exit with an error for a race condition
The daemon plugin pk-plugin-systemd-update deletes the /var/lib/PackageKit/prepared-update
file if any transaction is done that affects the prepared state.
This includes the offline-update process itself.

If the daemon deletes the file before the pk-offline-update process tries to do
it a warning is shown and the systemd unit fails.