lib: Actually return the error if any PkClient methods failed
Giovanni changed the g_new0() to a local helper object in commit
771f87943d3fc0a8579eee42ade373052208c8e2, but nothing actually set the return
value to NULL for failure. This meant we returned random memory for the
PkResults on error, and blew up with a segfault as soon as we tried accessing
the private structure.

It seems we could rely on the structure being zeroed by coincidence for small
numbers of packages, but for larger number of packages we got junk.