pisi: Fix parsing of package update messages
Many packages in SolusOS now use markdown in the package update sections,
to allow more informative messages to be shown in gpk-update-viewer.
Multiline messages that use markdown bullet lists are broken due to incorrect
handling of \n characters (replace with ";").

This commit fixes the issue by replacing new lines with the ";" character