lib: Do not have duplicate 'interactive' properties on the super and base classes
If the client called pk_client_set_interactive() on a PkTask instance this was
not being used in the derived class and the EULA and GPG auto-accept prompts
were failing.

There's no need at all to have the same property in both objects, the property
means the same in both. Deprecate the getters and setters in the superclass.

Fixes: https://bugzilla.gnome.org/show_bug.cgi?id=709676