hawkey: Emit percentage events when processing a large transaction
The global share check never got passed down and the all events were being
ignored, even though the small steps eventually added up to a large step. If it
turns out this is visible in the profiler, then we need to just fix the
percentage callbacks to be more efficient.