Remove the PkProvidesEnum parameter from WhatProvides
Most users of this just used ANY as the enum kind, and the backends were forced
to do many more string matches than actually required. It was a dumb idea to try
and codify all the different kinds of provides, as each time we had to wait for
PK to release and backends to be updated before we could make use of any of the
new types.

No soname bump as we have not had a tarball release since we last bumped.