gtk-doc: add missing setup_files.
Requires a yet to be commited patch to gtk-doc to generate
the subdirectories in the builddir (the setup files are copied
to the builddir from the sourcedir, with the relative path
kept, except the subdir of the relative path are not created if
missing as of now).

Depends on https://bugzilla.gnome.org/show_bug.cgi?id=720061 from
gtk-doc.

Signed-off-by: Richard Hughes <richard@hughsie.com>