Remove the package cache plugin
This was never satisfactory, and has been disabled by default for some time.
The reasons it's being removed are:

* Even when disabled, it still sets up the job which causes idle emmissions
* There was no way telling if the cache was up to date, or in anyway coherent
* We were adding a layer of caching on the existing database, when the real fix
  is to just fix the existing system database
* Parallel backends expode in a ball of fire when trying to execute plugin jobs
* Signals emitted from the plugin relied on there being a GMainContext
* The percentage completion went from 0..100 and then went back to 0..100 again

If we do want a package cache (and I don't think we actually do) then it should
be added into the backends directly, possibly making pk-package-cache.c
available to use in the backends. This would allow the parallelisation to work
correctly, and would also fix the percentage completion problem.