Revert "Make "reboot" the default action for no action file"
As a side effect, this commit inadvertently changed the default value
for the TriggerAction DBus property. This led to a fallout in the
gnome-shell shutdown dialog where the default then flipped to always
installing pending software updates on reboot / shutdown.

As for the issue that the original commit tried to address: it did so in
the wrong place. If there's a problem with the offline update process
hanging when no default is set, that should be addressed in
client/pk-offline-update.c to properly handle the UNSET state and not by
changing the default in the library.

This reverts commit 0a7b6d67b98cd06cabe03176c02db83f276dbe7b.