hif: Include any packages marked for installation when doing upgrades
When hawkey has depsolved an upgrade goal, make sure to not only read
back packages marked for update, but also the packages that are marked
for install. This ensures that we don't accidentally exclude a kernel
update just because hawkey has marked it for install, not update.

Also, since we're now including installable packages, this commit
removes a check that verified that all updated packages had an older
version already installed. Depending on a situation, hawkey might want
to install new packages during an update.

This is the other half needed for fixing
https://bugzilla.redhat.com/show_bug.cgi?id=1205649