pk-offline-update: exit quietly if we didn't prepare the update
To cooperate with other things using system-update.target, the
recommended behavior is to check to see if we prepared the update,
and if not, exit quietly so as to let the other updater run.

(http://lists.freedesktop.org/archives/systemd-devel/2015-July/033623.html)

This patch makes pk-offline-update check for the action file and the
prepared update transaction; if those don't exist, exit successfully.

Resolves: rhbz#1252500