offline update: Process updates triggered by older PackageKit versions
This commit adds compatibility with the previous PackageKit versions
that used to symlink /system-update to /var/cache.

When combined with online updates, we might get in a situation where the
offline updater has to process updates prepared by an older PackageKit.
In that case, if we simply 'exit 0', we end up in a system that always
boots in the offline update mode and hangs there.