cnf: minor fixes in command_not_found_handle{,r}() for bash and zsh
If pk-command-not-found is not run:
- Let command_not_found_handle() output to stderr, as bash would do.
- Use printf to output instead of echo to avoid issues with
  backslashes, also omit ': ' if $1 is empty or not set.
- Don't let command_not_found_handler() output anything, zsh does this
  for us.

This file was partly 'set -u' safe, now it's completly 'set -u' safe.

Ensure that command_not_found_handler() always returns false as it
should - you never know which future changes happen in the wrapped
command_not_found_handle().

The reason for the different name and the different behaviour is that
the people on the zsh-workers list did not want to copy a name they
considered to be wrong.  What they did copy from bash (possibly from a
downstream) was the initial suboptimal behaviour of
command_not_found_handle(), which later was changed to provide a more
flexible interface.