pkcon: Fix obsoleting packages to show up
Remove an old workaround which prevented replacement packages that
obsolete others to show up in pkcon.

This makes it go from:

$ pkcon update
The following packages have to be obsoleted:
 fedup-0.9.2-2.fc23.noarch	The Fedora Upgrade tool
Proceed with changes? [N/y]

to:

$ pkcon update
The following packages have to be obsoleted:
 fedup-0.9.2-2.fc23.noarch	The Fedora Upgrade tool
The following packages have to be installed:
 dnf-plugin-system-upgrade-0.7.1-1.fc22.noarch	System Upgrade plugin for DNF
 python2-dnf-plugin-system-upgrade-0.7.1-1.fc22.noarch	System Upgrade plugin for DNF
Proceed with changes? [N/y]