hif: Add support for UpgradeSystem
This hooks up the per-job HifContext changes that the previous commits
prepared.

Normally, we would have one global HifContext that is used for the
currently running distro. However, when we start processing a distro
upgrade, we create a new HifContext and use it only for the currently
running job.

This makes sure that anything that deals with the currently running
distro keeps using the cached context, but when dealing with infrequent
distro upgrades, we create a throwaway new context each time.