zypp: Enforce formatting input as string
A number of calls were implicitly handling data as strings,
which can be manipulated in unforeseen ways. It also broke
the build when -Werror=format-security was included
in the compiler flags (as it is by default on Fedora).

This is now changed to explicitly declare the input is a string.