Fix GIR annotations for progress callbacks
Otherwise, this makes Python applications crash when they try to run one
of the async functions.
While "notified" is still not the right type for this, it is the only
type which allows multiple function calls and survival of the main
function call.
The scopes "call" (lifetime only for function call) and "async" (lives
only until called once) are both wrong for this case.