backend job: Avoid emitting duplicates with parallel download
If we have a backend that supports parallel package downloads such as
the hif backend, we might end up alternating between different package
downloads in rapid succession. This can quickly overwhelm the message
bus.

This commit changes pk_backend_job_package to avoid emitting changes if
we've already emitted that exact one in the past; this should keep the
behaviour exactly the same for backends that do sequential downloads but
cut down emitted changes for backends that do parallel downloads.

https://bugzilla.redhat.com/show_bug.cgi?id=1336404#c20