fix updates and change code to aid in not breaking this again (#149)
* aptcc: fix resolution of updates by restoring previous behavior

before the local files resolution change everything that wasn't remove or
install was treated as install, this implicitly included updates and was
lost as part of the change.
reinstate updates being handled as installed to fix updates actually
working.

* aptcc: remove obscene fallback mechanic on name resolution

when not fixing broken installations we do a resolution on removes,
installs, updates, and local files (which are resolved differently) to
map them into actual packages we can work with.
this has an implicit component in that only remove and install were
explicitly handled and everything else was caught by an else branch
resolving everything that wasn't install/remove, like an install.

doing it this way had two major disadvantages:
a) we can't tell apart when the resolution failed (i.e. empty list
   returned) versus when we simply can't handle the requested role leading
   to utterly useless error messages
b) since it isn't explicitly defined what is handled in which way it's easy
   to forget about actual handling that is happening in the else

explicit handling makes the code more expressive and allows better error
reporting.