aptcc: make toinstall more apt compliant
new arguments: autoInst and preserveAuto

apt's markInstall can automatically try to resolve dependencies (incl.
recommends if applicable) without having the resolver deal with broken
dependencies at all. to that end it will mark dependencies of $package as
auto-install which gives them a slightly lower resolution score and the
auto flag (so they can be autoremoved).
To facilitate its use, the interface now calls toinstall twice: once
without autoInst to only mark explicitly wanted packages, and a second time
with autoInst to then also mark all needed dependencies. This is
specifically meant to allow or-relationships to resolve as desired.

e.g.
```
Package: foobar
Depends: foo|bar
```
with `pkcon install foobar bar` will explicitly mark foobar and bar and
thus prevent foo from getting autoinstalled.

In update scenarios it is hard to tell whether a user wants a package to be
installed or updated. In particular whether or not a package is meant to
be installed according to the user would prevent them from being
autoremoved. A user may however run `pkcon update foo` to update foo, not
because she actually thinks foo is needed or useful. To prevent
unintentionally losing the auto flag on updates, the backend now
always preserves the flag when updates are requested. Conversely on
installs we default to not having the auto flag.

fixes #153
fixes #151
fixes #145