aptcc: exclude held packages from update list
We can not accurately represent apt's held concept in PackageKit so to
mimic the intended behavior we simply exclude held packages from the
update listing. `pkcon update` and friends will not call update on held
packages, making them functionally held at their current version.
This has the added benefit that manually calling update
on the package will, just like install, break the version hold as the user
had to explicitly request the update (similar to apt's behavior on this).

fixed #120