dnf: Add locking for every threaded func
We are seeing a number of crashes being reported that look like they are
related to threading. This commit adds a big hammer mutex guard around
every threaded func. Let's see if this helps with the crashes; if it
does there is likely something wrong with the PK job scheduler that is
calling the backend (a non-parallelized backend) from multiple threads
concurrently.