aptcc: Fix a crash when the user supplies a bad package ID
Through pkcon the user may supply package IDs directly. When we then look
for them through apt-pkg's FindGrp we may get a group iterator that is
not actually valid. The PackageList obtained from that group iterator
can contain random garbage so as a pre-condition to doing anything with the
packagelist we need to check if the group at hand is good.
If the group is not good any packages it may list or not will by default be
useless, so we should not iterate on them.

For the record: on apt 1.2.19 I actually have FindGrp come back as bad
but then give out a PackageList where the first iterator is both good and
not at the end, which seems a bit meh from an API behavior point of view
but somewhat within reason given the owning iterator (i.e. the group)
itself is in a bad state.