pk-progress-bar: checks NULL strings inside pk_progress_bar_start ().
This commit adds a conditional to check if the string of progress bar
and text input are NULL. The comparison needs to be executed with two
valid strings and not two NULL pointes. That's why the conditional has
an AND operator. Both needs to be an empty string or a valid string, not
a NULL pointer. This missing check is causing an error during the pkcon
execution. See the command output:

$ sudo pkcon repair
FIXME: need to call pk_progress_bar_start()
earlier![=========================]
Finished                      [=========================]
                              [=========================]
Testing changes               [=========================]
Finished                      [=========================]

After patching:

$ sudo ./pkcon repair
                              [=========================]
Finished                      [=========================]
                              [=========================]
Testing changes               [=========================]
Finished                      [=========================]

Signed-off-by: Julio Faracco <jfaracco@br.ibm.com>