pk-main: segfault when config file is not found [v3].
When the config file used by PackageKit daemon is not found, the
function `pk_util_get_config_filename()` returns a NULL pointer.

See:
https://github.com/hughsie/PackageKit/blob/master/src/pk-shared.c#L300

So, when g_key_file_load_from_file() is called, the variable
`conf_filename` is NULL and inside that function the `assert()` fails:

    $ sudo packagekitd
    (packagekitd:9133): GLib-CRITICAL **: g_key_file_load_from_file:
    assertion 'file != NULL' failed
    Segmentation fault

This commits fixes this problem.

Signed-off-by: Julio Faracco <jfaracco@br.ibm.com>