aptcc: Don't return installed packages for codec search
It doesn't seem possible to perform this search properly with the
information we have. In particular we almost always say that '-good' is
a candidate when it might not be. '-good' is almost always installed, so
it is good enough to filter out installed packages.

The argument as to why this is okay to do is that if the user is being
shown the dialog then it's because they don't have a codec installed, so
it's not that helpful to show installed packages.