aptcc: fix offline updates by sanitizing origin more aggressively
origin may contain any number of nonesense, strip everything but words and
digits out!
specifically comma is breaking offline updates as the list of 'prepared'
packages is joined using comma, so having lingering commas in our package
ids utterly breaks splitting of the list and thus the offline update as a
whole.
seeing as origin is entirely free form it is probably wise to be very
aggressive in what we remove.

I've also added a bunch more details to the formats involved here for
prosperity and future reference.