dnf: Make sure cache invalidation doesn't destroy a DnfSack we are using
Keep a ref on DnfSack while we are using it; this makes sure that if we
happen to get cache invalidation from another thread, it doesn't destroy
the DnfSack underneath us.

It's unclear if it's an actual problem or not, but we do have a number
of weird crashes in libdnf code in bugzilla that could possibly be
explained by this. In any case this change makes the ownership semantics
easier to follow when reading the code.