aptcc: fix RequireRestart emission by moving it out of destructor
previously we would try to send RequireRestart in the dtor of AptIntf. this
emission was rejected by pk (supposedly because the transaction is
finished by that time already).
to fix this the RequireRestart was moved to surround the actual
installation/removal of packages making the signal sent before the
transaction is finished but after all packages are finished.

Fixes #234