dnf: Clean up leftover cache directories
After a distro upgrade to a new Fedora version, we always switch to a
new cache directory (it has the release version in the name). This
however so far meant that nothing cleaned up old cache directories from
previous releases.

This commit tries to address this and adds logic to clean up old cache
directories every time packagekitd starts up. We are also going to need
to find a good way to clean up the current cache directory, but that's a
separate fix for the future.

This also bumps glib dep to 2.54 for g_ascii_string_to_unsigned.

https://bugzilla.redhat.com/show_bug.cgi?id=1306992