aptcc: add support for downgrades as part of updates
users and vendors can "pin" a specific package version after the fact which
allows a specific version to be forced instead whatever the policy would
pick automatically. this sometimes constitutes a downgrade from an already
installed version. within an apt case these downgrades are resolved as
part of full-upgrades, equally within packagekit we'd inform the frontend
of downgrades via package emission. do this when get_updates is run.

the actual upgrade component of this is already implemented for quite a
while. the only thing that was missing is actually notifying the frontend
of the available downgrade.