dnf: Don't abort() when failing to set up context
If we abort() in init, then there's no way to report the error to
gnome-software.

Instead, log a warning when failing to set up context in init, and keep
going. Later, when one of the vfuncs is invoked and we have no context,
try setting it up again -- at this point, if it fails, we're able to
return the error message to gnome-software, so it can show it to the
user.