aptcc: Not all downloads have to be packages
When the mirror method is used in apt 1.6 or newer, a mirror list is
downloaded prior to any package, causing PackageKit to crash after
casting its item to pkgAcqArchiveSane.

pkgAcqArchiveSane is a local workaround to expose some protected
stuff in pkgAcqArchive, so we first do a dynamic_cast to pkgAcqArchive
which will be nullptr if it is not a pkgAcqArchive and then cast it
statically to pkgAcqArchiveSane.

Bug-Ubuntu: https://launchpad.net/bugs/1754265