Shutdown the daemon on idle by default
This reverts commit c6eb3555ec5b41e988c111d276764d55fb83bda3 and means that
the daemon will close after 5 minutes of inactivity.

The percieved race on shutdown is mitigated by the engine unown-ing the D-Bus
name in finalize. For machines with 2GB of RAM unloading the SAT database is a
big deal.

The slight downside here is that doing operations on the command line with
dnf or rpm will not trigger an updates_changed signal in GNOME Software if
PackageKit is not running, but the changed event should still be sent when the
the inotify watches of things in the AsStore fire.