main: Fix critical on idle shutdown
When doing the idle shutdown, the shutdown callback removes the shutdown
idle source from the main loop. Move things around a bit so that this
doesn't trigger a critical when cleaning up the same source at the end
of main().

(packagekitd:9749): GLib-CRITICAL **: 23:51:30.619: Source ID 13 was not found when attempting to remove it