dnf: Fix multithreaded race and memory corruption when reloading repos
When we get a file monitor callback from DnfRepoLoader, it runs in the
main thread and can race with a worker thread. This can lead to memory
corruption if we reload the repos and invalidate context's priv->repos
from the main thread and a worker thread is currently accessing them.

To fix it, this commit moves the repo reload out of the file monitor
callback and into the worker threads, avoiding the race.

We could also use locking, but then that would block the main daemon
thread, which is not ideal.