aptcc: Disable distro-upgrade functionality (for now)
This feature is broken currently and causes users problems who
accidentally trigger it somehow (it never worked on Debian, and
currently doesn't on Ubuntu as well).
We should actually fix this (maybe by using AppStream directly to fetch
release information?), but that implementation will require some
redesign and time.
So, disable this for now so users don't accidentally trigger it and get
shown weird errors.