aptcc: Do not emit PK_ERROR_ENUM_GPG_FAILURE
This prevents an apparently infinite loop due to automatic
retrying. We cannot be sure whether we had a GPG error and/or
other errors, so raising an error that causes automatic
retrying is a bad idea.

This will cause PK_ERROR_ENUM_CANNOT_FETCH_SOURCES to
be emitted by the caller of the function.