zypp: Reset upgrade mode to FALSE when not needed
Previously, after upgrade mode is set to TRUE, it’ll stay to be TRUE
util PackageKit daemon quits(it’s configured to quit after idling for
15 seconds). This can lead to some unexpected error somehow.

Fix that by spawning a timeout function after setting it to TRUE. We
cannot direct reset it, because updates-packages relies upgrade mode
in Tumbleweed to perform distribution upgrade. 1 second seems to be
enough for update-packages to be called, and then we can cancel the
timeout function and reset upgrade mode after all actions are done in
update-packages.