dnf: Load all the repos and vars directories
Historically, the backend has internally determined its setup with
static values. However, we generally want PackageKit to load all
repositories defined in all repository directories that DNF normally
searches, since it is not guaranteed to be in /etc/yum.repos.d and
DNF supports multiple repository configuration paths.

We also need the vars to be loaded so that repository definitions
that rely on more than the built-in vars will work.

This bumps our dependency for libdnf to 0.43.1, as we're using APIs
introduced in this release. This required bumping the CI environment
to Fedora 31.